Get this to compile again.  I think the -LANG:std addition broke it.
git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler_1.0@2464 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ test/threads/test.cpp
@@ -95,7 +95,7 @@ void doFunc(void* TSCounterPtr)
    while((*ts_counter) < 1000)
    {
       for(i=0;i<1000;i++)
-         float x = sin(i);
+         float x = sin((float) i);
       (*ts_counter)++;
    }
